Implement XInputTabletManager and provide an ANT script for UNIX.
Note that this code requires a modification to the JPen source
before it works. Specifically, the both of the following classes
must be marked as "public":

jpen.provider.xinput.XiDevice
jpen.provider.xinput.XiBus

Also note that this code still has quite a bit ahead of it.
The JTablet demo applet recognizes the tablet in its Screen
Listener tab, but for some reason doesn't actually draw
with it...