A bunch of changes to the build scripts.
Includes:
  * Changes to build directory structure
  * Changes to targets
  * Removal of targets (notably jtablet-tiny)
  * Fixing of indents
  * Etc.

Likely the most important change (besides the loss of jtablet-tiny)
is a change in the semantics of dist.version. It now contains the
exact version, complete with build date and svn.version. The dist
actual dist directory is now ${dir.dist}/${dist.version}.

Quite a few (perhaps too many) references to impl.version were
changed to dist.version in this commit as well. This is just
warning that build scripts may now start unexpectedly returning the
complete version number (not just 1.2.5-alpha)